requirements.txt
google-cloud-storage
langchain-google-community
langchain-text-splitters
langchain-community
langgraph
langchain[openai]
langchain-postgres
unstructured[pdf]
streamlit
python-dotenv
psycopg[binary]
python-docx
PyPDF2